tests: add check that 'Add shared filesystem' is not available for session connection guests
createVm helper function had to be extended to support session
connection VMs.

Note: because machines tests are non-destructive we need to make sure
that all resources created by each test are deleted in the teardown of
it. For system connection that's pretty straightforward, as all XMLs are
stored in /etc/libvirt/qemu directory, which we delete, and after
restarting the daemon everything is fresh new.

For the session connection XMLs are stored in
/home/admin/.cache/libvirt/ this does not work the same way.
Just just killing session libvirtd daemon after deleting the directory
still does not convince libvirt about the removal of the resources.

So let's manually undefine the pools and networks which get in the way
of the test non-destructiveness.